# apps/frontend/openapi/openapi.json
{
  "components": {
    "schemas": {
      "AggregatedValidationResult": {
        "description": "Combined results from multiple validation layers.",
        "example": {
          "execution_time_ms": 13.5,
          "layers_validated": [
            "airport_icao",
            "tac_syntax",
            "xml_wellformed"
          ],
          "passed": false,
          "results": [
            {
              "execution_time_ms": 5.2,
              "issues": [],
              "layer": "airport_icao",
              "passed": true
            },
            {
              "execution_time_ms": 8.3,
              "issues": [
                {
                  "code": "INVALID_TAC_FORMAT",
                  "layer": "tac_syntax",
                  "level": "error",
                  "location": "line 1, column 18",
                  "message": "Invalid TAC format: missing wind speed",
                  "suggestion": "Check wind speed component"
                }
              ],
              "layer": "tac_syntax",
              "passed": false
            }
          ],
          "total_issues": 2,
          "validated_at": "2026-02-10T14:30:45.123456"
        },
        "properties": {
          "execution_time_ms": {
            "default": 0,
            "minimum": 0.0,
            "title": "Execution Time Ms",
            "type": "number"
          },
          "layers_validated": {
            "description": "Layers validated",
            "items": {
              "$ref": "#/components/schemas/src__schemas__validation__ValidationLayer"
            },
            "title": "Layers Validated",
            "type": "array"
          },
          "passed": {
            "description": "Whether all layers passed",
            "title": "Passed",
            "type": "boolean"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/ValidationResult"
            },
            "title": "Results",
            "type": "array"
          },
          "total_issues": {
            "default": 0,
            "description": "Total issues",
            "minimum": 0.0,
            "title": "Total Issues",
            "type": "integer"
          },
          "validated_at": {
            "format": "date-time",
            "title": "Validated At",
            "type": "string"
          }
        },
        "required": [
          "passed",
          "layers_validated"
        ],
        "title": "AggregatedValidationResult",
        "type": "object"
      },
...
      "BatchValidationRequest": {
        "description": "Request to validate multiple inputs.",
        "example": {
          "items": [
            {
              "content": "METAR KJFK 231751Z 18012KT 10SM FEW040 15/07 A3005",
              "content_type": "tac"
            },
            {
              "content": "METAR EGLL 231750Z 17008KT 9999 SCT035 12/08 Q1007",
              "content_type": "tac"
            }
          ],
          "layers": [
            "airport_icao",
            "tac_syntax"
          ]
        },
        "properties": {
          "items": {
            "description": "Items to validate",
            "items": {
              "$ref": "#/components/schemas/ValidationRequest"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Items",
            "type": "array"
          },
          "layers": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationLayer-Input"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Layers to apply to all items (None = all layers)",
            "title": "Layers"
          }
        },
        "required": [
          "items"
        ],
        "title": "BatchValidationRequest",
        "type": "object"
      },
...
      "Body_validate_comprehensive_api_v1_validate_post": {
        "properties": {
          "exchange_profile": {
            "default": "",
            "description": "Exchange packaging profile (e.g. GLOBAL_AFS); ignored on validate-only paths",
            "title": "Exchange Profile",
            "type": "string"
          },
          "extensions": {
            "default": [],
            "description": "Optional national extension tokens (e.g. IWXXM_CA for full Canadian validate stack)",
            "items": {
              "type": "string"
            },
            "title": "Extensions",
            "type": "array"
          },
          "iwxxm_version": {
            "default": "2025-2",
            "description": "Target IWXXM version",
            "title": "Iwxxm Version",
            "type": "string"
          },
          "layers": {
            "default": [
              "ALL"
            ],
            "description": "Validation layers to run (ALL, or specific: AIRPORT_ICAO, TAC_SYNTAX, XML_WELLFORMED, XML_SCHEMA, SCHEMATRON, GML_REFERENCES, WMO_CODELISTS)",
            "items": {
              "type": "string"
            },
            "title": "Layers",
            "type": "array"
          },
          "manual_text": {
            "default": "",
            "description": "METAR TAC text to validate",
            "title": "Manual Text",
            "type": "string"
          },
          "product": {
            "default": "METAR",
            "description": "TAC product for Canadian extension XSD when extensions include IWXXM_CA",
            "title": "Product",
            "type": "string"
          },
          "profile": {
            "default": "",
            "description": "Deprecated - use semantic_profile (legacy alias: annex3 or iwxxm_us)",
            "title": "Profile",
            "type": "string"
          },
          "request_body": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ValidateRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "semantic_profile": {
            "default": "",
            "description": "Semantic profile id (e.g. ICAO_2025, US_FAA_NWS, CA_ECCC, AU_BOM, NZ_CAA_MET, UK_METOFFICE; aliases annex3 / iwxxm_us accepted)",
            "title": "Semantic Profile",
            "type": "string"
          },
          "stop_on_error": {
            "default": true,
            "description": "Stop at first blocking layer failure",
            "title": "Stop On Error",
            "type": "boolean"
          },
          "xml_content": {
            "default": "",
            "description": "Optional XML to validate (if omitted, TAC will be converted first)",
            "title": "Xml Content",
            "type": "string"
          }
        },
        "title": "Body_validate_comprehensive_api_v1_validate_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidateResponse": {
        "description": "Response for POST /api/v1/validate (validation layers + package_* extras).",
        "properties": {
          "extensions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Resolved national extension tokens from the request (when supplied)",
            "title": "Extensions"
          },
          "is_valid": {
            "title": "Is Valid",
            "type": "boolean"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ValidateIssueModel"
            },
            "title": "Issues",
            "type": "array"
          },
          "issues_by_layer": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/ValidateLayerIssueModel"
              },
              "type": "array"
            },
            "title": "Issues By Layer",
            "type": "object"
          },
          "layers_failed": {
            "items": {
              "type": "string"
            },
            "title": "Layers Failed",
            "type": "array"
          },
          "layers_passed": {
            "items": {
              "type": "string"
            },
            "title": "Layers Passed",
            "type": "array"
          },
          "layers_run": {
            "items": {
              "type": "string"
            },
            "title": "Layers Run",
            "type": "array"
          },
          "package_issues": {
            "items": {
              "$ref": "#/components/schemas/PackageIssueModel"
            },
            "title": "Package Issues",
            "type": "array"
          },
          "package_ok": {
            "default": true,
            "title": "Package Ok",
            "type": "boolean"
          },
          "package_stages": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PackageStageModel"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional per-stage breakdown from iwxxm-validate when profile=ca_eccc and extensions include IWXXM_CA",
            "title": "Package Stages"
          },
          "profile": {
            "default": "annex3",
            "title": "Profile",
            "type": "string"
          },
          "segments": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DecodeSegmentModel"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional item-by-item decode rows (code and explanation) when a readable decode exists. Omitted when there is no decode.",
            "title": "Segments"
          },
          "stopped_at_layer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stopped At Layer"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional plain-language paragraph of the decoded report when a readable decode exists. Omitted when there is no decode.",
            "title": "Summary"
          },
          "total_issues": {
            "default": 0,
            "title": "Total Issues",
            "type": "integer"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "is_valid",
          "version"
        ],
        "title": "ValidateResponse",
        "type": "object"
      },
...
      "ValidationRequest": {
        "description": "Request for validation operation.",
        "example": {
          "content": "METAR KJFK 231751Z 18012KT 10SM FEW040 15/07 A3005 RMK AO2",
          "content_type": "tac",
          "iwxxm_version": "3.0.1",
          "layers": [
            "airport_icao",
            "tac_syntax"
          ]
        },
        "properties": {
          "content": {
            "description": "Content to validate (METAR TAC or IWXXM XML)",
            "examples": [
              "METAR KJFK 231751Z 18012KT 10SM FEW040 15/07 A3005 RMK AO2"
            ],
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "content_type": {
            "default": "tac",
            "description": "Content type: 'tac' (METAR TAC) or 'xml' (IWXXM XML)",
            "examples": [
              "tac",
              "xml"
            ],
            "title": "Content Type",
            "type": "string"
          },
          "iwxxm_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "IWXXM version for validation context",
            "examples": [
              "3.0.1"
            ],
            "title": "Iwxxm Version"
          },
          "layers": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationLayer-Input"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Specific layers to validate (None = all layers)",
            "title": "Layers"
          }
        },
        "required": [
          "content"
        ],
        "title": "ValidationRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Convert METAR/SPECI TAC messages to IWXXM XML format with comprehensive validation",
    "title": "METAR to IWXXM Backend API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/validate": {
      "post": {
        "description": "Perform comprehensive 7-layer IWXXM validation.\n\nValidates METAR TAC input through all 7 validation layers:\n\n1. **Layer 1 (AIRPORT_ICAO)**: Validates ICAO airport code against database\n2. **Layer 2 (TAC_SYNTAX)**: Validates TAC/METAR syntax basics\n3. **Layer 3 (XML_WELLFORMED)**: Checks XML is well-formed\n4. **Layer 4 (XML_SCHEMA)**: Validates against official IWXXM XSD schemas\n5. **Layer 5 (SCHEMATRON)**: Validates business rules from official Schematron\n6. **Layer 6 (GML_REFERENCES)**: Validates GML internal references\n7. **Layer 7 (WMO_CODELISTS)**: Validates against official WMO RDF codelists\n\n**Authentication**: Public (no login required)\n\n**Request Parameters**:\n- **manual_text** (required): METAR TAC text to validate\n- **xml_content** (optional): Pre-converted XML to validate (if omitted, TAC will be converted)\n- **iwxxm_version**: Target IWXXM version (default: \"2025-2\")\n- **layers**: Validation layers to run (default: [\"ALL\"])\n  - \"ALL\": Run all 7 layers\n  - Or specify: [\"AIRPORT_ICAO\", \"TAC_SYNTAX\", \"XML_SCHEMA\", \"SCHEMATRON\", ...]\n- **stop_on_error**: Stop at first blocking layer failure (default: true)\n\n**Response**:\n```json\n{\n  \"is_valid\": true,\n  \"version\": \"2025-2\",\n  \"layers_run\": [\"AIRPORT_ICAO\", \"TAC_SYNTAX\", \"XML_WELLFORMED\", \"XML_SCHEMA\", \"SCHEMATRON\", \"GML_REFERENCES\", \"WMO_CODELISTS\"],\n  \"layers_passed\": [\"AIRPORT_ICAO\", \"TAC_SYNTAX\", \"XML_WELLFORMED\", \"XML_SCHEMA\", \"SCHEMATRON\", \"GML_REFERENCES\", \"WMO_CODELISTS\"],\n  \"layers_failed\": [],\n  \"total_issues\": 0,\n  \"issues_by_layer\": {},\n  \"stopped_at_layer\": null\n}\n```",
        "operationId": "validate_comprehensive_api_v1_validate_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_validate_comprehensive_api_v1_validate_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Comprehensive",
        "tags": [
          "Validation",
          "Validation"
        ]
      }
    },
...
    "/api/v1/validation/validate": {
      "post": {
        "description": "Validate METAR TAC or IWXXM XML content through multiple validation layers.\n\n## Request Body\n- **content** (string, required): The METAR TAC or IWXXM XML content to validate\n- **content_type** (string, default=\"tac\"): Type of content (\"tac\" or \"xml\")\n- **layers** (array, optional): Specific validation layers to run. If None, runs all layers:\n  - `airport_icao`: Validate ICAO airport code\n  - `tac_syntax`: Validate METAR TAC syntax\n  - `xml_wellformed`: Check XML is well-formed\n  - `xml_schema`: Validate against XSD schema\n  - `schematron`: SCHEMATRON rules validation\n  - `gml_references`: GML reference checks\n  - `wmo_codelists`: WMO code list validation\n- **iwxxm_version** (string, optional): IWXXM version for context (e.g., \"3.0.1\")\n\n## Response\nReturns aggregated validation results with:\n- **passed** (boolean): Whether all requested layers passed\n- **layers_validated** (array): Layers that were run\n- **total_issues** (integer): Total validation issues found\n- **results** (array): Per-layer validation details\n- **execution_time_ms** (float): Total execution time\n\n## Example Success Response\n```json\n{\n  \"passed\": true,\n  \"layers_validated\": [\"airport_icao\", \"tac_syntax\"],\n  \"total_issues\": 0,\n  \"results\": [\n    {\n      \"passed\": true,\n      \"layer\": \"airport_icao\",\n      \"issues\": [],\n      \"execution_time_ms\": 5.2\n    }\n  ],\n  \"execution_time_ms\": 10.5\n}\n```\n\n## Example Failure Response\n```json\n{\n  \"passed\": false,\n  \"layers_validated\": [\"airport_icao\", \"tac_syntax\"],\n  \"total_issues\": 1,\n  \"results\": [\n    {\n      \"passed\": false,\n      \"layer\": \"tac_syntax\",\n      \"issues\": [\n        {\n          \"layer\": \"tac_syntax\",\n          \"level\": \"error\",\n          \"message\": \"Invalid TAC format\",\n          \"code\": \"INVALID_TAC_FORMAT\"\n        }\n      ],\n      \"execution_time_ms\": 8.3\n    }\n  ],\n  \"execution_time_ms\": 8.3\n}\n```",
        "operationId": "validate_content_api_v1_validation_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AggregatedValidationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate METAR TAC or IWXXM XML content",
        "tags": [
          "Validation",
          "Validation"
        ]
      }
    },
    "/api/v1/validation/validate-multi": {
      "post": {
        "description": "Validate multiple METAR TAC or IWXXM XML inputs in a single request.\n\nUseful for batch validation of multiple entries. Each item is validated\nindependently and can have different content types.\n\n## Request Body\n- **items** (array, required): Array of validation requests (1-100 items)\n  - Each item has: content, content_type, layers (optional), iwxxm_version (optional)\n- **layers** (array, optional): Default layers to apply to all items\n\n## Response\nReturns batch validation results with:\n- **results** (array): Aggregated validation result for each item\n- **total_items** (integer): Total items requested\n- **passed_items** (integer): Items that passed all layers\n- **failed_items** (integer): Items that failed at least one layer\n- **total_execution_time_ms** (float): Total execution time for all items\n\n## Example Request\n```json\n{\n  \"items\": [\n    {\n      \"content\": \"METAR KJFK 231751Z 18012KT 10SM FEW040 15/07 A3005\",\n      \"content_type\": \"tac\"\n    },\n    {\n      \"content\": \"METAR EGLL 231750Z 17008KT 9999 SCT035 12/08 Q1007\",\n      \"content_type\": \"tac\"\n    }\n  ],\n  \"layers\": [\"airport_icao\", \"tac_syntax\"]\n}\n```\n\n## Example Response\n```json\n{\n  \"results\": [\n    {\n      \"passed\": true,\n      \"layers_validated\": [\"airport_icao\", \"tac_syntax\"],\n      \"total_issues\": 0,\n      \"results\": [],\n      \"execution_time_ms\": 12.3\n    },\n    {\n      \"passed\": false,\n      \"layers_validated\": [\"airport_icao\", \"tac_syntax\"],\n      \"total_issues\": 1,\n      \"results\": [],\n      \"execution_time_ms\": 8.5\n    }\n  ],\n  \"total_items\": 2,\n  \"passed_items\": 1,\n  \"failed_items\": 1,\n  \"total_execution_time_ms\": 20.8\n}\n```",
        "operationId": "validate_multiple_api_v1_validation_validate_multi_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate multiple METAR TAC or IWXXM XML inputs",
        "tags": [
          "Validation",
          "Validation"
        ]
      }
    },
...
  },
...
}